+++ src/store.js
@@ -63,7 +63,7 @@ Project*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.store = '2023-June-05';
+modules.store = '2023-June-06';
 
 // XML_Serializer ///////////////////////////////////////////////////////
 /*
@@ -1060,7 +1060,8 @@ SnapSerializer.prototype.loadCustomBlocks = function (
                                 : child.contents,
                         options ? options.contents : undefined,
                         child.attributes.readonly === 'true',
-                        child.attributes.irreplaceable === 'true'
+                        child.attributes.irreplaceable === 'true',
+                        child.attributes.separator
                     ]
                 );
             });
@@ -2321,12 +2322,17 @@ CustomBlockDefinition.prototype.toXML = function (serializer) {
         Array.from(this.declarations.keys()).reduce((xml, decl) => {
             // to be refactored now that we've moved to ES6 Map:
                 return xml + serializer.format(
-                    '<input type="@"$$>$%</input>',
+                    '<input type="@"$$$>$%</input>',
                     this.declarations.get(decl)[0],
                     this.declarations.get(decl)[3] ?
                             ' readonly="true"' : '',
                     this.declarations.get(decl)[4] ?
                             ' irreplaceable="true"' : '',
+                    this.declarations.get(decl)[5] ?
+                            ' separator="' +
+                                this.declarations.get(decl)[5] +
+                                '"'
+                            : '',
                     this.declarations.get(decl)[1],
                     this.declarations.get(decl)[2] ?
                             serializer.format(