+++ HISTORY.md
@@ -60,7 +60,8 @@
 ### 2023-06-06
 * blocks: added "with" infix to JOIN, made it collapse from 2 inputs to input list, omitting the single input case
 * German translation update for "with" (as JOIN infix label)
-* byob, blocks: new option to specify a "separator" (infix) label for variadic input slots inside custom blocks 
+* byob, blocks: new option to specify a "separator" (infix) label for variadic input slots inside custom blocks
+* blocks, byob, store: serialize variadic input slot separators
 
 ### 2023-06-05
 * blocks: made drop-halo for variadic slots white (instead of red) in "explicit input list" mode